requirements.txt
torch
torchvision
fastapi
uvicorn
pillow
numpy
opencv-python-headless
python-multipart